newThoughtAbove
